lisa.trace: Remove unused event sanitization
If needed, this processing will be reintroduced as part of an analysis,
as the sanitization is only there to provide values similar to what the
kernel manipulates.

Column names modifications or arbitrary augmentation of the data needs
to be done in separate functions, with custom parameters driving their
behavior, and with the ability to deprecate them if necessary.